refactor: relocate components used by ApCalcPage and HomePage
