--- src/client/components/ApCalc/index.js
@@ -1,5 +1,0 @@
-// this folder houses what used to be children of APCalcCard,
-// which has been turned into a standalone page.
-import Results from "./Results";
-import FormField from "./FormField";
-export { Results, FormField };
