+++ src/pages/HomePage/components/ChaldeaGateCard/ScheduleRow.jsx
