+++ src/pages/HomePage/components/ChaldeaGateCard/TrainingLootList.jsx
