+++ src/pages/HomePage/components/LoginTicketCard/index.js
