+++ src/pages/HomePage/components/ShopCard/ShopCard.jsx
