+++ src/pages/HomePage/components/ShopCard/ShopCardListItem.jsx
